add sorting module and move default logic there
