+++ corehq/apps/userreports/reports/specs.py
@@ -1,4 +1,5 @@
 from django.utils.translation import ugettext as _
+from corehq.apps.userreports.reports.sorting import ASCENDING, DESCENDING
 from dimagi.ext.jsonobject import JsonObject, StringProperty, BooleanProperty, ListProperty, DictProperty, ObjectProperty
 from jsonobject.base import DefaultProperty
 from sqlagg import CountUniqueColumn, SumColumn
@@ -24,8 +25,6 @@ SQLAGG_COLUMN_MAP = {
     'simple': SimpleColumn,
     'year': YearColumn,
 }
-ASCENDING = "ASC"
-DESCENDING = "DESC"
 
 
 class ReportFilter(JsonObject):
